WIP Student (Activity) Groups
WIP - I didnt have groups for lecture/lessons, now finally Im implementing them.

+++ src/main/java/com/studencki/TimePlan/dtos/ActivityDTO.java
@@ -13,5 +13,5 @@ public class ActivityDTO {
     private int duration;
     private int type;
 
-    private Integer groupNumber;
+    private Long groupNumber;
 }
